Review feedback to remove old project, move bridgeable extensions to designated file
