--- IntegratingSDK/IntegratingSDK/IntegratingSDK.h
@@ -1,19 +1,0 @@
-//
-//  IntegratingSDK.h
-//  IntegratingSDK
-//
-//  Created by Joe Susnick on 4/3/19.
-//  Copyright © 2019 facebook. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for IntegratingSDK.
-FOUNDATION_EXPORT double IntegratingSDKVersionNumber;
-
-//! Project version string for IntegratingSDK.
-FOUNDATION_EXPORT const unsigned char IntegratingSDKVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <IntegratingSDK/PublicHeader.h>
-
-
