--- Samples/ObjCCatalog/ObjCCatalog/ObjCCatalog/Restaurant.h
@@ -1,19 +1,0 @@
-//
-//  Restaurant.h
-//  IntegratingSDK
-//
-//  Created by Joe Susnick on 4/3/19.
-//  Copyright © 2019 facebook. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@import FacebookCore;
-
-@interface Restaurant : NSObject
-
-@end
-
-NS_ASSUME_NONNULL_END
